+++ SBBDesignSystemMobileSwiftUIDemoTests/ViewModifiers/SBBSwitchStyleTests.swift
@@ -2,18 +2,18 @@
 // Copyright (C) Schweizerische Bundesbahnen SBB, 2021.
 //
 
-import XCTest
+import Testing
 import SnapshotTesting
 import SwiftUI
 import SBBDesignSystemMobileSwiftUI
 
-class SBBSwitchStyleTests: XCTestCase {
+class SBBSwitchStyleTests: DSMTest {
     
     var toggle = Toggle(isOn: .constant(true)) {
         Text("Label")
     }
 
-    func testSwitchStyle() {
+    @Test func testSwitchStyle() {
         let view = toggle.toggleStyle(SBBSwitchStyle())
         for colorScheme in ColorScheme.allCases {
             view.recordDocumentationSnapshot(name: "SBBSwitchStyle", colorScheme: colorScheme)
@@ -21,7 +21,7 @@ class SBBSwitchStyleTests: XCTestCase {
         }
     }
     
-    func testSwitchStyleDisabled() {
+    @Test func testSwitchStyleDisabled() {
         let view = toggle.toggleStyle(SBBSwitchStyle()).disabled(true)
         for colorScheme in ColorScheme.allCases {
             assertSnapshot(matching: view.colorScheme(colorScheme).toVC(), as: imagePortrait, record: recordNewReferenceSnapshots)