+++ SBBDesignSystemMobileSwiftUIDemoTests/Views/SBBInfoViewTests.swift
@@ -2,16 +2,16 @@
 // Copyright (C) Schweizerische Bundesbahnen SBB, 2021.
 //
 
-import XCTest
+import Testing
 import SnapshotTesting
 import SwiftUI
 import SBBDesignSystemMobileSwiftUI
 
-class SBBInfoViewTests: XCTestCase {
+class SBBInfoViewTests: DSMTest {
     
     var text = Text("In und um den Bahnhof Biel / Bienne wird viel gebaut. Das kann Auswirkungen auf Ihre Reise haben. Beachten Sie deshalb den Onlinefahrplan und die aktuellen Anzeigen am Bahnhof, um über geänderte Gleise und Fahrpläne informiert zu sein.")
 
-    func testInfoView() {
+    @Test func testInfoView() {
         let view = SBBInfoView(image: Image(sbbIcon: .station_small), text: text)
         for colorScheme in ColorScheme.allCases {
             view.frame(width: 375).recordDocumentationSnapshot(name: "SBBInfoView", colorScheme: colorScheme)